firewall/requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.30.0
httpx==0.28.1
pyyaml==6.0.3
dependency-injector==4.48.2
sentence-transformers==5.1.2
onnxruntime==1.23.2
qdrant-client==1.15.1
structlog==25.5.0